# telemetry: get rid of the tracing_fmt layer and use our own optimized one