--- a/apollo-router/Cargo.toml
+++ b/apollo-router/Cargo.toml
@@ -228,7 +228,7 @@ tracing = "0.1.37"
 tracing-core = "0.1.31"
 tracing-futures = { version = "0.2.5", features = ["futures-03"] }
 tracing-opentelemetry = "0.21.0"
-tracing-subscriber = { version = "0.3.17", features = ["env-filter", "json"] }
+tracing-subscriber = { version = "0.3.18", features = ["env-filter", "json"] }
 trust-dns-resolver = "0.23.2"
 url = { version = "2.4.1", features = ["serde"] }
 urlencoding = "2.1.3"