--- a/apollo-router/src/plugins/telemetry/config_new/logging.rs
+++ b/apollo-router/src/plugins/telemetry/config_new/logging.rs
@@ -328,6 +328,8 @@ impl Default for JsonFormat {
 #[derive(Deserialize, JsonSchema, Clone, Debug, Eq, PartialEq)]
 #[serde(deny_unknown_fields, rename_all = "snake_case", default)]
 pub(crate) struct TextFormat {
+    /// Enable ansi (default: true)
+    pub(crate) ansi: bool,
     /// Include the timestamp with the log event.
     pub(crate) display_timestamp: bool,
     /// Include the target with the log event.
@@ -353,6 +355,7 @@ pub(crate) struct TextFormat {
 impl Default for TextFormat {
     fn default() -> Self {
         TextFormat {
+            ansi: true,
             display_timestamp: true,
             display_target: false,
             display_level: true,