… Cargo.lock
… apollo-router/src/configuration/snapshots/apollo_router__configuration__tests__schema_generation.snap
… apollo-router/src/configuration/snapshots/apollo_router__configuration__tests__upgrade_old_configuration@telemetry-exporters.router.yaml.snap
… apollo-router/src/plugins/telemetry/formatters/text.rs
… apollo-router/src/plugins/telemetry/reload.rs