Some perf improvements to lazy parse the attribute for logs

```
NEW ===
Summary:
  Total:        78.7140 secs
  Slowest:      0.1418 secs
  Fastest:      0.0055 secs
  Average:      0.0196 secs
  Requests/sec: 2540.8439


OLD ===
Summary:
  Total:        86.7263 secs
  Slowest:      0.1005 secs
  Fastest:      0.0065 secs
  Average:      0.0216 secs
  Requests/sec: 2306.1057
```

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR is marked ready-for-review.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention of other (impacted) groups. Please endeavour to do this before seeking PR approval. The mechanism for doing this will vary considerably, so use your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding Manual Tests, please document the manual testing (extensively) in the Exceptions.