@bnjjj 2023-11-15
@garypen Yes I will add tests now, but in a follow up PR to make sure PRs stay small and easy to review. But it will be easier to test our logs now we have our own layer I think